+++ misskey-api/src/endpoint/i/registry.rs
@@ -1,0 +1,8 @@
+pub mod get;
+pub mod get_all;
+pub mod get_detail;
+pub mod keys;
+pub mod keys_with_type;
+pub mod remove;
+pub mod scopes;
+pub mod set;