+++ misskey-api/src/model.rs
@@ -37,6 +37,7 @@ pub mod note_reaction;
 pub mod notification;
 pub mod page;
 pub mod query;
+pub mod registry;
 pub mod signin;
 pub mod sort;
 pub mod user;